EvgenProdTools: fix to allow for long int event number, GeneratorFilters: make...

See merge request atlas/athena!57285

+++ Generators/EvgenProdTools/src/CountHepMC.cxx
@@ -59,7 +59,7 @@ StatusCode CountHepMC::execute() {
   ATH_MSG_DEBUG("Current count = " << m_nPass);
   ATH_MSG_INFO("Options for HepMC event number, EvtID event number, EvtID run number = " << m_corHepMC << m_corEvtID << m_corRunNumber );
   // Fix the event number
-  int newnum = m_nPass + m_firstEv - 1;
+  long long int newnum = m_nPass + m_firstEv - 1;
   if (newnum<=0){
     ATH_MSG_ERROR("Event number must be positive-definite; " << newnum << " is not allowed");
     return StatusCode::FAILURE;
